+++ src/migrations/0002_messages_table.py
@@ -20,7 +20,7 @@ def execute(obj):
     # database settings
     table_name = 'messages'
     rename_columns = [('timestamp', 'updated_at')]
-    add_columns = [('created_at', 'TIMESTAMP', 'CURRENT_TIMESTAMP')]
+    add_columns = [('created_at', 'TIMESTAMP', 'CURRENT_TIMESTAMP'), ('state', 'VARCHAR(255)', 'added')]
 
     # check if the table exists and has the necessary schema for execute the migration
     # check table
@@ -46,10 +46,10 @@ def execute(obj):
             obj.database_connection.rollback()
 
         print(f"{NAME}: Columns in the {table_name} table have been renamed")
-        try:
-            for column in add_columns:
+        for column in add_columns:
+            try:
                 obj.cursor.execute(f"ALTER TABLE {table_name} ADD COLUMN {column[0]} {column[1]} DEFAULT {column[2]}")
                 obj.database_connection.commit()
-        except obj.errors as error:
-            print(f"{NAME}: Columns in the {table_name} table have not been added. Skip adding: {error}")
-            obj.database_connection.rollback()
+            except obj.errors as error:
+                print(f"{NAME}: Columns in the {table_name} table have not been added. Skip adding: {error}")
+                obj.database_connection.rollback()